Move Data to rely less on shared examples
