+++ core/data/to_s_spec.rb
@@ -1,6 +1,9 @@
 require_relative '../../spec_helper'
-require_relative 'shared/inspect'
+require_relative 'fixtures/classes'
 
 describe "Data#to_s" do
-  it_behaves_like :data_inspect, :to_s
+  it "is an alias of Data#inspect" do
+    a = DataSpecs::Measure.new(42, "km")
+    a.method(:to_s).should == a.method(:inspect)
+  end
 end
